% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_proxy_bdf.R
\name{proxy_bdf}
\alias{proxy_bdf}
\title{Configuration of the BDF Proxy server}
\usage{
proxy_bdf()
}
\description{
========================================================
=====  THIS FUNCTION IS FOR BDF INTERNAL USE ONLY  =====
========================================================
}
\details{
(It should not do anything if accidentally used outside of the BDF environment.)

Cette fonction permet de configurer R pour se connecter au proxy BDF.
Il est nécessaire de rentrer son mot de passe INTRA.
Il faut lancer cette fonction à chaque session.
}
\keyword{internal}
