% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{w_meta}
\alias{w_meta}
\title{Get metadata from a dataframe returned by w_data function}
\usage{
w_meta(data)
}
\arguments{
\item{data}{dataframe from w_data()}
}
\value{
dataframe
}
\description{
Get metadata from a dataframe returned by w_data function
}
\examples{
\dontrun{
## Request the US Dollar monthly exchange rates in Euro
df <- w_data(dataset_name = "EXR", series_name = "M.USD.EUR.SP00.E")
meta <- w_meta(df)
}

}
